requirements.txt
pandas
scikit-learn
snowflake-connector-python
joblib